added styling for chat pages and refactored tests to use helper modules
