+++ app/assets/stylesheets/messages.scss
@@ -1,3 +1,39 @@
 // Place all the styles related to the Messages controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: http://sass-lang.com/
+.chat-messages {
+  display:flex;
+  flex-direction: column;
+  align-items: center;
+}
+.chat-messages ul {
+  list-style: none;
+  padding: 0;
+  max-width: 90%;
+}
+.chat-messages li {
+  display: -webkit-flex;
+  display: flex;
+  -webkit-flex-direction: column;
+  flex-direction: column;
+  max-width: 75%;
+  background-color: #ffffff;
+  -webkit-box-shadow: 0 3px 8px rgba(0,0,0,0.25);
+  padding: 8px;
+  margin: auto;
+  margin-bottom: 15px;
+  border-radius: 10px;
+}
+.chat-message-form textarea {
+  -webkit-box-shadow: 0 3px 8px rgba(0,0,0,0.25);
+  border-radius: 10px;
+  border: none;
+  width: 100%;
+  height: 4em;
+}
+.chat-message-form input {
+  -webkit-box-shadow: 0 3px 8px rgba(0,0,0,0.25);
+  border-radius: 10px;
+  border: none;
+  background: #e3e7ed;
+}
