+++ app/controllers/chats_controller.rb
@@ -7,11 +7,10 @@ class ChatsController < ApplicationController
     @sender = current_user
     @recipient = User.find_by(email: chat_params[:email])
     if Chat.between(@sender.id, @recipient.id).present?
-      @chat = Chat.between(@sender.id, @recipient.id)
+      @chat = Chat.between(@sender.id, @recipient.id).first
     else
       @chat = Chat.create({sender_id: @sender.id, recipient_id: @recipient.id})
     end
-
     redirect_to chat_messages_path(@chat)
   end
