+++ app/views/chats/new.html.erb
@@ -1,15 +1,16 @@
 <h1>Chat</h1>
 
-
+<section class="new-chat-form">
 <%= form_for :chat, url: chats_path do |f| %>
   <p>
     <%= f.label :email %>
   </p>
   <p>
-    <%= f.text_field :email %>
+    <%= f.text_field :email,  :required => true %>
   </p>
   <p>
-    <%= f.submit %>
+    <%= f.submit "Start a chat" %>
   </p>
 
 <% end %>
+</section>
