+++ spec/spec_helper.rb
@@ -1,4 +1,7 @@
 require_relative "./support/make_post"
+require_relative "./support/user_sessions"
+require_relative "./support/chat_helpers"
+
 # This file was generated by the `rails generate rspec:install` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
 # The generated `.rspec` file contains `--require spec_helper` which will cause
@@ -19,6 +22,9 @@ require_relative "./support/make_post"
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
   config.include PostHelpers, :type => :feature
+  config.include UserHelpers, :type => :feature
+  config.include ChatHelpers, :type => :feature
+
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
   # assertions if you prefer.
